sbd: introduce basic framework for Sheepdog Block Device
This is similar to Ceph's RBD. The main motivation is to replace complex
and ineffecient middle ware (such as iscci softwafe) with simple software stacks
to expose sheepdog storage as Linux block device interface.

Usage:

We control the device the same way as RBD.

# associate vdi 'test' to /dev/sbd0
$ echo 127.0.0.1 7000 test > /sys/bus/sbd/add

# remove the device sbd0
$ echo 0 > /sys/bus/sbd/remove

For now we don't do actual read/write yet, just a aio framework sketched out.

Signed-off-by: Liu Yuan <namei.unix@gmail.com>